adjusting wick grammar, renaming wick.js to wick_parser.js in import statements
