+++ source/compiler/grammars/html.hcg
@@ -80,8 +80,7 @@
 
 <> TEXT_NODE → BODY_SYMBOLS(+") 							↦cstr ^text
 
-<> BODY_SYMBOLS → θid │ θws │ θsym │ θop │ θob │ θcb │ θnum │ θkeyword │ \(
-
+<> BODY_SYMBOLS → θid │ θws │ θsym │ θop │ θob │ θcb │ θnum │ θkeyword
