+++ source/compiler/grammars/string.hcg
@@ -6,4 +6,4 @@
 
 <> string_data → ( string_data_val )(+") 	
 
-<> string_data_val → θid │ θws │ θany │ θsym │ θop │ θcb │ θob
+<> string_data_val → θid │ θws │ θany │ θsym │ θop │ θcb │ θob
