+++ source/compiler/html/element.js
@@ -1,7 +1,7 @@
 import { appendChild, createElement } from "../../short_names.js";
 import Scope from "../component/runtime/scope.js";
 import ComponentEnvironment from "../component_environment.js";
-import wick_compile from "../wick.js";
+import wick_compile from "../wick_parser.js";
 import error from "../../utils/error.js";
 import URL from "@candlefw/url";
 import whind from "@candlefw/whind";
